fix: resolve draft-pod infinite recursion and pool-gen u32 overflow
- Fix infinite React re-render on /draft-pod route caused by Zustand
  selectors returning new [] refs when view is null (HostControls,
  SeatStatusRing)
- Widen sheet weights from u32 to u64 in draft-core to handle MTGJSON
  sets (TLA, FIN) with totalWeight/card weights exceeding u32::MAX
- Add draft crate sources to Tiltfile wasm deps and new draft-pools
  data resource

+++ crates/draft-core/src/extraction.rs
@@ -58,8 +58,8 @@ struct MtgjsonBoosterPlay {
 #[derive(Deserialize)]
 #[serde(rename_all = "camelCase")]
 struct MtgjsonSheet {
-    cards: HashMap<String, u32>,
-    total_weight: u32,
+    cards: HashMap<String, u64>,
+    total_weight: u64,
     #[serde(default)]
     foil: bool,
     #[serde(default)]
